Added doc comments, renamed some example files and classes. Added torch::tril()
